requirements.txt
scipy==1.10.1
PyPDF2
Flask==3.1.2
pdfplumber==0.11.5
python-docx==1.1.2
numpy==1.26.4
scikit-learn==1.3.2